// src/app/main/interfaces/actualiza-datos.interface.ts
export interface Actualiza {
  nombre_delegacion:  string;
  codigo_postal:      number;
  domicilio:          string;
  coordinador:        string;
  coordinador_puesto: string;
  coordinador_genero: string;
  secretario:         string;
  secretario_puesto:  string;
  secretario_genero:  string;
}